Added instructions (musall-mice) on constructing videos from svd data
