+++ scripts/musall-mice/data-preprocessing/download_data.sh
